Refactor main.py for OOP and Enhance Logs Robustness
Refactored the procedural code in the main.py file into a class-based
structure, improving modularity and readability. Extracted core
application setup and launch steps into `created_envioment` and `launch`
methods within the new `NH0` class, encapsulating the initialization
process. Additionally, updated the Logs class in utils/logs.py to check
for the existence of a notification item before deletion, preventing
potential errors during runtime.

These changes lay the groundwork for more object-oriented development
practices and enhance application stability by avoiding operations on
non-existent GUI elements.

+++ utils/logs.py
@@ -104,7 +104,8 @@ class Logs:
 
     def close_notification(self, notification: int | str, delay: float = 5.0):
         time.sleep(delay)
-        dpg.delete_item(notification)
+        if dpg.does_item_exist(notification):
+            dpg.delete_item(notification)
 
     # Defining a method to log a message with a specific format type
     def log(self, message: str, f_type: str = "s"):
@@ -169,6 +170,7 @@ class Logs:
         """
         self.df.to_csv(self.csv_file_path, mode="w")
         print(f"Logs saved into {self.csv_file_path}")
+
 
 def verify_singleton():
     temp: Logs = Logs.getInstance()
